refactored OVA classes to ova_utils folder
