--- k8_vmware/vsphere/OVA.py
@@ -1,278 +1,0 @@
-import atexit
-import os
-import os.path
-import ssl
-import sys
-import tarfile
-import time
-
-from threading import Timer
-from argparse import ArgumentParser
-from getpass import getpass
-
-import pyVmomi
-from osbot_utils.utils import Http
-from osbot_utils.utils.Files import file_exists, file_not_exists
-from six.moves.urllib.request import Request, urlopen
-
-from k8_vmware.vsphere.Sdk import Sdk
-
-
-class OVA:
-
-    def __init__(self): # todo add methods to manipulated OVA files
-        self.sdk = Sdk()
-
-    def upload_ova(self, ova_path):
-        sdk = self.sdk
-        # assert file_exists(ova_path)
-        # todo: add explicit check for file exists and return error
-        # todo: normalise method return values
-        ovf_handle = OvfHandler(ova_path)
-
-        ovfManager = sdk.content().ovfManager
-        rp = sdk.resource_pool()
-        ds = sdk.datastore()
-        dc = sdk.datacenter()
-        cisp = pyVmomi.vim.OvfManager.CreateImportSpecParams()
-        cisr = ovfManager.CreateImportSpec(ovf_handle.get_descriptor(), rp, ds, cisp)
-        ovf_handle.set_spec(cisr)
-
-        lease = rp.ImportVApp(cisr.importSpec, dc.vmFolder)
-        while lease.state == pyVmomi.vim.HttpNfcLease.State.initializing:
-            print("Waiting for lease to be ready...")
-            time.sleep(1)
-
-        if lease.state == pyVmomi.vim.HttpNfcLease.State.error:
-            print("Lease error: %s" % lease.error)
-            return 1
-
-        if lease.state == pyVmomi.vim.HttpNfcLease.State.done:
-            return 0
-
-        host = sdk.server_details().get('host')
-
-        print("Starting deploy...")
-        result = ovf_handle.upload_disks(lease, host)
-        print(result)
-
-    def download_ova_file(self, url, target_ova_path):
-        if file_not_exists(target_ova_path):
-            Http.GET_bytes_to_file(url, target_ova_path)
-        return target_ova_path
-
-def get_tarfile_size(tarfile):
-    """
-    Determine the size of a file inside the tarball.
-    If the object has a size attribute, use that. Otherwise seek to the end
-    and report that.
-    """
-    if hasattr(tarfile, 'size'):
-        return tarfile.size
-    size = tarfile.seek(0, 2)
-    tarfile.seek(0, 0)
-    return size
-
-class OvfHandler(object):      #todo: Create sepeate class
-    """
-    OvfHandler handles most of the OVA operations.
-    It processes the tarfile, matches disk keys to files and
-    uploads the disks, while keeping the progress up to date for the lease.
-    """
-    def __init__(self, ovafile):
-        """
-        Performs necessary initialization, opening the OVA file,
-        processing the files and reading the embedded ovf file.
-        """
-        self.handle = self._create_file_handle(ovafile)
-        self.tarfile = tarfile.open(fileobj=self.handle)
-        ovffilename = list(filter(lambda x: x.endswith(".ovf"),
-                                  self.tarfile.getnames()))[0]
-        ovffile = self.tarfile.extractfile(ovffilename)
-        self.descriptor = ovffile.read().decode()
-
-    def _create_file_handle(self, entry):
-        """
-        A simple mechanism to pick whether the file is local or not.
-        This is not very robust.
-        """
-        if os.path.exists(entry):
-            return FileHandle(entry)
-        else:
-            return WebHandle(entry)
-
-    def get_descriptor(self):
-        return self.descriptor
-
-    def set_spec(self, spec):
-        """
-        The import spec is needed for later matching disks keys with
-        file names.
-        """
-        self.spec = spec
-
-    def get_disk(self, fileItem, lease):
-        """
-        Does translation for disk key to file name, returning a file handle.
-        """
-        ovffilename = list(filter(lambda x: x == fileItem.path,
-                                  self.tarfile.getnames()))[0]
-        return self.tarfile.extractfile(ovffilename)
-
-    def get_device_url(self, fileItem, lease):
-        for deviceUrl in lease.info.deviceUrl:
-            if deviceUrl.importKey == fileItem.deviceId:
-                return deviceUrl
-        raise Exception("Failed to find deviceUrl for file %s" % fileItem.path)
-
-    def upload_disks(self, lease, host):
-        """
-        Uploads all the disks, with a progress keep-alive.
-        """
-        self.lease = lease
-        try:
-            self.start_timer()
-            for fileItem in self.spec.fileItem:
-                self.upload_disk(fileItem, lease, host)
-            lease.Complete()
-            print("Finished deploy successfully.")
-            return 0
-        except pyVmomi.vmodl.MethodFault as e:
-            print("Hit an error in upload: %s" % e)
-            lease.Abort(e)
-        except Exception as e:
-            print("Lease: %s" % lease.info)
-            print("Hit an error in upload: %s" % e)
-            lease.Abort(pyVmomi.vmodl.fault.SystemError(reason=str(e)))
-            raise
-        return 1
-
-    def upload_disk(self, fileItem, lease, host):
-        """
-        Upload an individual disk. Passes the file handle of the
-        disk directly to the urlopen request.
-        """
-        ovffile = self.get_disk(fileItem, lease)
-        if ovffile is None:
-            return
-        deviceUrl = self.get_device_url(fileItem, lease)
-        url = deviceUrl.url.replace('*', host)
-        headers = {'Content-length': get_tarfile_size(ovffile)}
-        if hasattr(ssl, '_create_unverified_context'):
-            sslContext = ssl._create_unverified_context()
-        else:
-            sslContext = None
-        req = Request(url, ovffile, headers)
-        urlopen(req, context=sslContext)
-
-    def start_timer(self):
-        """
-        A simple way to keep updating progress while the disks are transferred.
-        """
-        Timer(5, self.timer).start()
-
-    def timer(self):
-        """
-        Update the progress and reschedule the timer if not complete.
-        """
-        try:
-            prog = self.handle.progress()
-            self.lease.Progress(prog)
-            if self.lease.state not in [pyVmomi.vim.HttpNfcLease.State.done,
-                                        pyVmomi.vim.HttpNfcLease.State.error]:
-                self.start_timer()
-            sys.stderr.write("Progress: %d%%\r" % prog)
-        except:  # Any exception means we should stop updating progress.
-            pass
-
-
-class FileHandle(object):     #todo: Create sepeate class
-    def __init__(self, filename):
-        self.filename = filename
-        self.fh = open(filename, 'rb')
-
-        self.st_size = os.stat(filename).st_size
-        self.offset = 0
-
-    def __del__(self):
-        self.fh.close()
-
-    def tell(self):
-        return self.fh.tell()
-
-    def seek(self, offset, whence=0):
-        if whence == 0:
-            self.offset = offset
-        elif whence == 1:
-            self.offset += offset
-        elif whence == 2:
-            self.offset = self.st_size - offset
-
-        return self.fh.seek(offset, whence)
-
-    def seekable(self):
-        return True
-
-    def read(self, amount):
-        self.offset += amount
-        result = self.fh.read(amount)
-        return result
-
-    # A slightly more accurate percentage
-    def progress(self):
-        return int(100.0 * self.offset / self.st_size)
-
-
-class WebHandle(object):     #todo: Create sepeate class
-    def __init__(self, url):
-        self.url = url
-        r = urlopen(url)
-        if r.code != 200:
-            raise FileNotFoundError(url)
-        self.headers = self._headers_to_dict(r)
-        if 'accept-ranges' not in self.headers:
-            raise Exception("Site does not accept ranges")
-        self.st_size = int(self.headers['content-length'])
-        self.offset = 0
-
-    def _headers_to_dict(self, r):
-        result = {}
-        if hasattr(r, 'getheaders'):
-            for n, v in r.getheaders():
-                result[n.lower()] = v.strip()
-        else:
-            for line in r.info().headers:
-                if line.find(':') != -1:
-                    n, v = line.split(': ', 1)
-                    result[n.lower()] = v.strip()
-        return result
-
-    def tell(self):
-        return self.offset
-
-    def seek(self, offset, whence=0):
-        if whence == 0:
-            self.offset = offset
-        elif whence == 1:
-            self.offset += offset
-        elif whence == 2:
-            self.offset = self.st_size - offset
-        return self.offset
-
-    def seekable(self):
-        return True
-
-    def read(self, amount):
-        start = self.offset
-        end = self.offset + amount - 1
-        req = Request(self.url,
-                      headers={'Range': 'bytes=%d-%d' % (start, end)})
-        r = urlopen(req)
-        self.offset += amount
-        result = r.read(amount)
-        r.close()
-        return result
-
-    # A slightly more accurate percentage
-    def progress(self):
-        return int(100.0 * self.offset / self.st_size)
