+++ k8_vmware/vsphere/ova_utils/File_Handle.py
@@ -1,0 +1,38 @@
+import os
+
+
+class FileHandle(object):
+    def __init__(self, filename):
+        self.filename = filename
+        self.fh = open(filename, 'rb')
+
+        self.st_size = os.stat(filename).st_size
+        self.offset = 0
+
+    def __del__(self):
+        self.fh.close()
+
+    def tell(self):
+        return self.fh.tell()
+
+    def seek(self, offset, whence=0):
+        if whence == 0:
+            self.offset = offset
+        elif whence == 1:
+            self.offset += offset
+        elif whence == 2:
+            self.offset = self.st_size - offset
+
+        return self.fh.seek(offset, whence)
+
+    def seekable(self):
+        return True
+
+    def read(self, amount):
+        self.offset += amount
+        result = self.fh.read(amount)
+        return result
+
+    # A slightly more accurate percentage
+    def progress(self):
+        return int(100.0 * self.offset / self.st_size)
