+++ tests/unit/helpers/test_TestCase_OVA.py
@@ -4,6 +4,6 @@ class test_TestOVA_VM(TestCase_OVA):
 
     def test_setUpClass(self):
         assert type(self.vm_name   ) == str
-        assert self.vm_name is not None
-        assert self.url is not None
-        assert self.ova_path is not None
+        assert self.vm_name      is     not None
+        assert self.url          is     not None
+        assert self.ova_path     is     not None
