+++ tests/unit/vsphere/ova_utils/test_File_Handle.py
@@ -1,0 +1,46 @@
+from osbot_utils.utils.Files import file_exists
+
+from k8_vmware.helpers.TestCase_OVA import TestCase_OVA
+from k8_vmware.vsphere.ova_utils.File_Handle import FileHandle
+from k8_vmware.vsphere.ova_utils.OVA import OVA
+
+
+class test_file_handle(TestCase_OVA):
+    def setUp(self) -> None:
+        self.ova = OVA()
+        self.ova.download_ova_file(self.url, self.ova_path)
+        self.filehandle = FileHandle(self.ova_path)
+
+    def test_init(self):
+        assert file_exists(self.ova_path)
+        assert self.filehandle is not None
+
+    def test_tell(self):
+        response=self.filehandle.tell()
+        assert response is not None
+
+    def test_seekable(self):
+        response=self.filehandle.seekable()
+        assert response is True
+
+    def test_progress(self):
+        self.filehandle.offset=self.filehandle.st_size/2
+        response = self.filehandle.progress()
+        assert response is 50
+
+    def test_seek(self):
+        response=self.filehandle.seek(offset=1,whence=0)
+        assert response is 1
+
+        expected = self.filehandle.offset + 1
+        response=self.filehandle.seek(offset=1,whence=1)
+        assert response is expected
+
+        response = self.filehandle.seek(offset=1,whence=2)
+        assert response is not None
+
+    def test_read(self):
+        amount=10000
+        response=self.filehandle.read(amount=amount)
+        self.assertEqual( len(response.decode("utf-8",errors='ignore')), 10000)
+
